NP-3404 Move files
