+++ src/pages/search/SearchBar.tsx
@@ -1,19 +1,14 @@
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
-import { Box, Button, IconButton, MenuItem, TextField } from '@mui/material';
+import { Box, Button, IconButton, TextField } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
 import ClearIcon from '@mui/icons-material/Clear';
 import FilterAltIcon from '@mui/icons-material/FilterAltOutlined';
 import { Field, FieldArray, FieldArrayRenderProps, FieldProps, useFormikContext } from 'formik';
-import { dataTestId } from '../utils/dataTestIds';
-import { ExpressionStatement, PropertySearch, SearchConfig } from '../utils/searchHelpers';
-import {
-  ContributorFieldNames,
-  DescriptionFieldNames,
-  ResourceFieldNames,
-  SpecificContributorFieldNames,
-} from '../types/publicationFieldNames';
-import { BetaFunctionality } from './BetaFunctionality';
+import { dataTestId } from '../../utils/dataTestIds';
+import { ExpressionStatement, PropertySearch, SearchConfig } from '../../utils/searchHelpers';
+import { BetaFunctionality } from '../../components/BetaFunctionality';
+import { AdvancedSearchRow } from './filters/AdvancedSearchRow';
 
 const StyledTextField = styled(TextField)`
   margin-top: 0;
@@ -100,58 +95,5 @@ export const SearchBar = () => {
         </FieldArray>
       </BetaFunctionality>
     </>
-  );
-};
-
-interface AdvancedSearchRowProps {
-  baseFieldName: string;
-  removeFilter: () => void;
-}
-
-const AdvancedSearchRow = ({ removeFilter, baseFieldName }: AdvancedSearchRowProps) => {
-  const { t } = useTranslation('search');
-
-  return (
-    <Box sx={{ display: 'grid', gridTemplateColumns: '3fr 3fr 6fr 2fr', gap: '1rem' }}>
-      <Field name={`${baseFieldName}.fieldName`}>
-        {({ field }: FieldProps<string>) => (
-          <StyledTextField
-            {...field}
-            select
-            variant="outlined"
-            label={t('field_label')}
-            InputLabelProps={{ shrink: true }}>
-            <MenuItem value={DescriptionFieldNames.Title}>{t('common:title')}</MenuItem>
-            <MenuItem value={DescriptionFieldNames.Abstract}>{t('registration:description.abstract')}</MenuItem>
-            <MenuItem value={ResourceFieldNames.SubType}>{t('registration_type')}</MenuItem>
-            <MenuItem value={DescriptionFieldNames.Tags}>{t('registration:description.keywords')}</MenuItem>
-            <MenuItem value={`${ContributorFieldNames.Contributors}.${SpecificContributorFieldNames.Name}`}>
-              {t('registration:contributors.contributor')}
-            </MenuItem>
-          </StyledTextField>
-        )}
-      </Field>
-      <Field name={`${baseFieldName}.operator`}>
-        {({ field }: FieldProps<string>) => (
-          <StyledTextField {...field} select variant="outlined" label={t('operator')}>
-            <MenuItem value={ExpressionStatement.Contains}>{t('contains')}</MenuItem>
-            <MenuItem value={ExpressionStatement.NotContaining}>{t('not_containing')}</MenuItem>
-          </StyledTextField>
-        )}
-      </Field>
-      <Field name={`${baseFieldName}.value`}>
-        {({ field }: FieldProps<string>) => (
-          <StyledTextField
-            {...field}
-            variant="outlined"
-            label={t('search_term_label')}
-            InputLabelProps={{ shrink: true }}
-          />
-        )}
-      </Field>
-      <Button onClick={removeFilter} color="error">
-        {t('remove_filter')}
-      </Button>
-    </Box>
   );
 };
